+++ garnet3d/main/src/priv/inc/garnet/gfx/image.inl
@@ -4,8 +4,8 @@
 GN_INLINE bool GN::gfx::ImageDesc::setFaceAndLevel( size_t faces, size_t levels )
 {
     safeMemFree( mipmaps );
-    numFaces = faces;
-    numLevels = levels;
+    numFaces = (uint32_t)faces;
+    numLevels = (uint32_t)levels;
     if( numFaces > 0 && numLevels > 0 )
     {
         mipmaps = (MipmapDesc*)memAlloc( numFaces * numLevels * sizeof(MipmapDesc) );
